requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-compressor==2.4
django-libsass==0.8
django-render-partial==0.4
django-sass-processor==0.8.2
djangorestframework==3.12.4
libsass==0.20.1
npm==0.1.1
optional-django==0.1.0
Pillow==8.1.0
Pygments==2.8.1
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
